Issue 140: Investigate if and how we can extract the integration tests into suites
Implemented.
