+++ src/test/java/org/carlspring/cloud/storage/s3fs/AmazonS3ITSuite.java
@@ -1,4 +1,4 @@
-package org.carlspring.cloud.storage.s3fs.suites;
+package org.carlspring.cloud.storage.s3fs;
 
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.platform.suite.api.IncludeTags;
@@ -9,8 +9,8 @@ import org.junit.runner.RunWith;
  * @author carlspring
  */
 @RunWith(JUnitPlatform.class)
-@IncludeTags({ "s" })
-@SelectPackages("org.carlspring.cloud.storage.s3fs.path")
+@IncludeTags({ "s3-integration-test" })
+@SelectPackages({ "org.carlspring.cloud.storage.s3fs" })
 public class AmazonS3ITSuite
 {
 }
