+++ src/test/java/org/carlspring/cloud/storage/s3fs/FilesITTest.java
@@ -24,6 +24,7 @@ import java.util.UUID;
 
 import com.github.marschall.memoryfilesystem.MemoryFileSystemBuilder;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
@@ -37,7 +38,8 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class FilesIT
+@Tag("s3-integration-test")
+class FilesITTest
 {
 
     private static final String bucket = EnvironmentBuilder.getBucket();
