+++ src/test/java/org/carlspring/cloud/storage/s3fs/S3UtilsITTest.java
@@ -19,6 +19,7 @@ import java.util.UUID;
 
 import com.github.marschall.memoryfilesystem.MemoryFileSystemBuilder;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
@@ -33,8 +34,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-
-class S3UtilsIT
+@Tag("s3-integration-test")
+class S3UtilsITTest
 {
 
     private static final String bucket = EnvironmentBuilder.getBucket();
