+++ src/test/java/org/carlspring/cloud/storage/s3fs/fileSystemProvider/GetFileSystemITTest.java
@@ -11,6 +11,7 @@ import java.util.Properties;
 
 import com.google.common.collect.ImmutableMap;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import static org.carlspring.cloud.storage.s3fs.S3Factory.ACCESS_KEY;
 import static org.carlspring.cloud.storage.s3fs.S3Factory.REGION;
@@ -23,7 +24,8 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.spy;
 
-class GetFileSystemIT
+@Tag("s3-integration-test")
+class GetFileSystemITTest
 {
 
     private S3FileSystemProvider provider;
